Added Accuracy system, Fixed main menu
i added an accuracy system. fixed the main menu so now that tutorial scales better and that the credits button works. lowered hit box of the note goal so that the notes don't get delete too early and give a penalty. Also fixed the bug in the build that makes the speeeeeeeeen too much.